+++ lib/src/flutter/widget/scroll_controller.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/widgets.dart';
+
+extension ScrollControllerExtensions on ScrollController {
+  /// Verifies if scroll has reach the end of it.
+  bool get hasReachTheEndOfScroll {
+    return offset >= position.maxScrollExtent && !position.outOfRange;
+  }
+}