+++ lib/src/flutter/widget/widget.dart
@@ -1,0 +1,2 @@
+export 'responsivity.dart';
+export 'scroll_controller.dart';